+++ src/SQLProvider/Providers.MSAccess.fs
@@ -241,11 +241,14 @@ type internal MSAccessProvider(contextSchemaPath) =
                     |> Seq.map (fun row ->
                         match row.["DATA_TYPE"].ToString() |> findDbType with
                         |Some(m) ->
+                            let pkColumn = pks |> List.exists (fun idx -> idx = row.["COLUMN_NAME"].ToString())
                             let col =
                                 { Column.Name = row.["COLUMN_NAME"].ToString();
                                   TypeMapping = m
-                                  IsPrimaryKey = pks |> List.exists (fun idx -> idx = row.["COLUMN_NAME"].ToString())
+                                  IsPrimaryKey = pkColumn
                                   IsNullable = bool.Parse(row.["IS_NULLABLE"].ToString())
+                                  IsAutonumber = row.["DATA_TYPE"].ToString() = "AutoNumber"
+                                  HasDefault = not (row.IsNull "COLUMN_DEFAULT");
                                   TypeInfo = 
                                     try 
                                         let ti =