+++ src/SQLProvider/Providers.Oracle.fs
@@ -236,9 +236,15 @@ module internal Oracle =
         |> Seq.toList
 
     let getColumns (primaryKeys:IDictionary<_,_>) (table : Table) conn = 
-        sprintf """select data_type, nullable, column_name, data_length from all_tab_columns where table_name = '%s' and owner = '%s'""" table.Name table.Schema
+        sprintf """select data_type, nullable, column_name, data_length, data_default from all_tab_columns where table_name = '%s' and owner = '%s'""" table.Name table.Schema
         |> read conn (fun row ->
-                let columnType = Sql.dbUnbox row.[0]
+                let columnType : string = Sql.dbUnbox row.[0]
+                // Remove precision specification from the column type name (can appear in TIMESTAMP and INTERVAL)
+                // Example: 'TIMESTAMP(3) WITH TIMEZONE' must be transformed to 'TIMESTAMP WITH TIMEZONE'
+                let columnType = 
+                    match columnType.IndexOf('('), columnType.IndexOf(')') with
+                    | x,y when x > 0 && y > 0 -> columnType.Substring(0,x) + columnType.Substring(y+1)
+                    | _ -> columnType
                 let nullable   = (Sql.dbUnbox row.[1]) = "Y"
                 let columnName = Sql.dbUnbox row.[2]
                 let typeinfo = 
@@ -247,10 +253,13 @@ module internal Oracle =
                     else columnType + "(" + datalength + ")"
                 findDbType columnType
                 |> Option.map (fun m ->
+                    let pkColumn = primaryKeys.TryGetValue(table.Name) |> function | true, pks -> pks = [columnName] | false, _ -> false
                     { Name = columnName
                       TypeMapping = m
-                      IsPrimaryKey = primaryKeys.TryGetValue(table.Name) |> function | true, pks -> pks = [columnName] | false, _ -> false
+                      IsPrimaryKey = pkColumn
                       IsNullable = nullable
+                      IsAutonumber = pkColumn
+                      HasDefault = row.[4] <> null
                       TypeInfo = Some typeinfo }
                 ))
         |> Seq.choose id