+++ src/SQLProvider/SqlSchema.fs
@@ -42,6 +42,8 @@ type Column =
       TypeMapping: TypeMapping
       IsPrimaryKey: bool
       IsNullable: bool
+      IsAutonumber: bool
+      HasDefault: bool
       TypeInfo: string option }
     with
         static member FromQueryParameter(q: QueryParameter) =
@@ -49,6 +51,8 @@ type Column =
               TypeMapping = q.TypeMapping
               IsPrimaryKey = false
               IsNullable = true
+              IsAutonumber = false
+              HasDefault = false
               TypeInfo = None }
 
 type ColumnLookup = Map<string,Column>